chore: setup config
