+++ .env
@@ -1,0 +1,2 @@
+HTTP_PORT=8181
+WS_PORT=3000
